fix(http): links updated to match swagger doc
