+++ chronograf/ui/src/types/v2/links.ts
@@ -1,3 +1,6 @@
+// Links is the response from the /api/v2 endpoint.  It contains
+// links to all other endpoints.
+// see https://github.com/influxdata/platform/blob/db5b20f4eabcc7d2233e54415cbd48945b0b4d0c/http/api_handler.go#L125
 export interface Links {
   auths: string
   buckets: string
@@ -7,18 +10,21 @@ export interface Links {
   external: {
     statusFeed: string
   }
-  flux: {
-    ast: string
+  query: {
     self: string
+    ast: string
+    spec: string
     suggestions: string
   }
   orgs: string
-  query: string
   setup: string
+  signin: string
+  signout: string
   sources: string
   system: {debug: string; health: string; metrics: string}
   tasks: string
-  users: string
   write: string
+  users: string
+  macros: string
   defaultDashboard: string
 }
