+++ http/health.go
@@ -5,8 +5,8 @@ import (
 	"net/http"
 )
 
-// HealthzHandler returns the status of the process.
-func HealthzHandler(w http.ResponseWriter, r *http.Request) {
+// HealthHandler returns the status of the process.
+func HealthHandler(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(http.StatusOK)
 	fmt.Fprintln(w, `{"message": "howdy y'all", "status": "healthy"}`)
 }
