+++ http/query_handler.go
@@ -8,9 +8,13 @@ import (
 	"io"
 	"net/http"
 	"net/url"
+	"time"
 
 	"github.com/influxdata/flux"
+	"github.com/influxdata/flux/ast"
+	"github.com/influxdata/flux/complete"
 	"github.com/influxdata/flux/csv"
+	"github.com/influxdata/flux/parser"
 	"github.com/influxdata/platform"
 	pcontext "github.com/influxdata/platform/context"
 	"github.com/influxdata/platform/kit/errors"
@@ -30,6 +34,7 @@ type FluxHandler struct {
 
 	Logger *zap.Logger
 
+	Now                  func() time.Time
 	AuthorizationService platform.AuthorizationService
 	OrganizationService  platform.OrganizationService
 	ProxyQueryService    query.ProxyQueryService
@@ -39,10 +44,15 @@ type FluxHandler struct {
 func NewFluxHandler() *FluxHandler {
 	h := &FluxHandler{
 		Router: httprouter.New(),
+		Now:    time.Now,
 		Logger: zap.NewNop(),
 	}
 
 	h.HandlerFunc("POST", fluxPath, h.handlePostQuery)
+	h.HandlerFunc("POST", "/api/v2/query/ast", h.postFluxAST)
+	h.HandlerFunc("POST", "/api/v2/query/spec", h.postFluxSpec)
+	h.HandlerFunc("GET", "/api/v2/query/suggestions", h.getFluxSuggestions)
+	h.HandlerFunc("GET", "/api/v2/query/suggestions/:name", h.getFluxSuggestion)
 	return h
 }
 
@@ -89,6 +99,140 @@ func (h *FluxHandler) handlePostQuery(w http.ResponseWriter, r *http.Request) {
 			zap.String("handler", "flux"),
 			zap.Error(err),
 		)
+	}
+}
+
+type langRequest struct {
+	Query string `json:"query"`
+}
+
+type postFluxASTResponse struct {
+	AST *ast.Program `json:"ast"`
+}
+
+// postFluxAST returns a flux AST for provided flux string
+func (h *FluxHandler) postFluxAST(w http.ResponseWriter, r *http.Request) {
+	var request langRequest
+	ctx := r.Context()
+
+	err := json.NewDecoder(r.Body).Decode(&request)
+	if err != nil {
+		EncodeError(ctx, errors.MalformedDataf("invalid json: %v", err), w)
+		return
+	}
+
+	ast, err := parser.NewAST(request.Query)
+	if err != nil {
+		EncodeError(ctx, errors.InvalidDataf("invalid json: %v", err), w)
+		return
+	}
+
+	res := postFluxASTResponse{
+		AST: ast,
+	}
+
+	if err := encodeResponse(ctx, w, http.StatusOK, res); err != nil {
+		EncodeError(ctx, err, w)
+		return
+	}
+}
+
+type postFluxSpecResponse struct {
+	Spec *flux.Spec `json:"spec"`
+}
+
+// postFluxSpec returns a flux Spec for provided flux string
+func (h *FluxHandler) postFluxSpec(w http.ResponseWriter, r *http.Request) {
+	var req langRequest
+	ctx := r.Context()
+
+	err := json.NewDecoder(r.Body).Decode(&req)
+	if err != nil {
+		EncodeError(ctx, errors.MalformedDataf("invalid json: %v", err), w)
+		return
+	}
+
+	spec, err := flux.Compile(ctx, req.Query, h.Now())
+	if err != nil {
+		EncodeError(ctx, errors.InvalidDataf("invalid json: %v", err), w)
+		return
+	}
+
+	res := postFluxSpecResponse{
+		Spec: spec,
+	}
+
+	if err := encodeResponse(ctx, w, http.StatusOK, res); err != nil {
+		EncodeError(ctx, err, w)
+		return
+	}
+}
+
+// fluxParams contain flux funciton parameters as defined by the semantic graph
+type fluxParams map[string]string
+
+// suggestionResponse provides the parameters available for a given Flux function
+type suggestionResponse struct {
+	Name   string     `json:"name"`
+	Params fluxParams `json:"params"`
+}
+
+// suggestionsResponse provides a list of available Flux functions
+type suggestionsResponse struct {
+	Functions []suggestionResponse `json:"funcs"`
+}
+
+// getFluxSuggestions returns a list of available Flux functions for the Flux Builder
+func (h *FluxHandler) getFluxSuggestions(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	completer := complete.DefaultCompleter()
+	names := completer.FunctionNames()
+	var functions []suggestionResponse
+	for _, name := range names {
+		suggestion, err := completer.FunctionSuggestion(name)
+		if err != nil {
+			EncodeError(ctx, err, w)
+			return
+		}
+
+		filteredParams := make(fluxParams)
+		for key, value := range suggestion.Params {
+			if key == "table" {
+				continue
+			}
+
+			filteredParams[key] = value
+		}
+
+		functions = append(functions, suggestionResponse{
+			Name:   name,
+			Params: filteredParams,
+		})
+	}
+	res := suggestionsResponse{Functions: functions}
+
+	if err := encodeResponse(ctx, w, http.StatusOK, res); err != nil {
+		EncodeError(ctx, err, w)
+		return
+	}
+}
+
+// getFluxSuggestion returns the function parameters for the requested function
+func (h *FluxHandler) getFluxSuggestion(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	name := httprouter.ParamsFromContext(ctx).ByName("name")
+	completer := complete.DefaultCompleter()
+
+	suggestion, err := completer.FunctionSuggestion(name)
+	if err != nil {
+		EncodeError(ctx, err, w)
+		return
+	}
+
+	res := suggestionResponse{Name: name, Params: suggestion.Params}
+	if err := encodeResponse(ctx, w, http.StatusOK, res); err != nil {
+		EncodeError(ctx, err, w)
+		return
 	}
 }
